gdump: Fix some GError -Wsign-conversion warnings
Because of the generic nature of `GError`, `g_set_error()` has to take
an `int`, but `g_file_error_from_errno()` returns a `GFileError`. The
macOS CI runner decides that’s a good reason to emit
`-Wsign-conversion`.


Helps: #3405

+++ girepository/gdump.c
@@ -641,7 +641,7 @@ gi_repository_dump (const char  *input_filename,
   if (input == NULL)
     {
       int saved_errno = errno;
-      g_set_error (error, G_FILE_ERROR, g_file_error_from_errno (saved_errno),
+      g_set_error (error, G_FILE_ERROR, (int) g_file_error_from_errno (saved_errno),
                    "Failed to open ‘%s’: %s", input_filename, g_strerror (saved_errno));
 
       g_module_close (self);
@@ -653,7 +653,7 @@ gi_repository_dump (const char  *input_filename,
   if (output == NULL)
     {
       int saved_errno = errno;
-      g_set_error (error, G_FILE_ERROR, g_file_error_from_errno (saved_errno),
+      g_set_error (error, G_FILE_ERROR, (int) g_file_error_from_errno (saved_errno),
                    "Failed to open ‘%s’: %s", output_filename, g_strerror (saved_errno));
 
       fclose (input);
@@ -735,7 +735,7 @@ gi_repository_dump (const char  *input_filename,
       {
         int saved_errno = errno;
 
-        g_set_error (error, G_FILE_ERROR, g_file_error_from_errno (saved_errno),
+        g_set_error (error, G_FILE_ERROR, (int) g_file_error_from_errno (saved_errno),
                      "Error closing input file ‘%s’: %s", input_filename,
                      g_strerror (saved_errno));
         caught_error = TRUE;
@@ -745,7 +745,7 @@ gi_repository_dump (const char  *input_filename,
       {
         int saved_errno = errno;
 
-        g_set_error (error, G_FILE_ERROR, g_file_error_from_errno (saved_errno),
+        g_set_error (error, G_FILE_ERROR, (int) g_file_error_from_errno (saved_errno),
                      "Error closing output file ‘%s’: %s", output_filename,
                      g_strerror (saved_errno));
         caught_error = TRUE;
